requirements.txt
numpy~=1.24.4
opencv-python~=4.8.1.78
Pillow~=10.1.0
tqdm~=4.66.1
pycocotools~=2.0.7
torch-dreams==4.0.0
IQA-pytorch==0.1